test: split Wasm test cases into subdirectories
