--- e2e/cases/wasm/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@rsbuild/config/tsconfig",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "outDir": "./dist",
-    "paths": {
-      "@/*": ["wasm-basic/src/*"]
-    }
-  },
-  "include": ["wasm-basic/src", "*.test.ts", "index.test.ts"]
-}
